Add basic functionality
The script gets the current host's IP from icanhazip.com, validates that
it looks like a v4 IP (so we aren't inserting arbitrary untrusted text
into our configuration file), run it through a template to generate a
munin-node.conf file, and uploads that to the remote server. Then it
compares the new and old files, and if they are different it replaces
the old file and restarts `munin-node`.