Change to a script you can run on the remote machine
Instead of having the script ssh to a remote machine (requring a root
login), now you can install this command on the remote machine and grant
passwordless sudo access to it.

I added added remote-munin-node-access (just a shell script) which you
can run locally out of cron to get your ip and run the actual command on
the remote box(es).